SingularVectors

Documentation for SingularVectors.

SingularVectors.UniversalEnvelope.AbstractElem

SingularVectors.UniversalEnvelope.AbstractEnvElem

SingularVectors.UniversalEnvelope.LieElem

SingularVectors.UniversalEnvelope.LieEnvElem

SingularVectors.UniversalEnvelope.SupElem

SingularVectors.UniversalEnvelope.SupEnvElem

Data types

Element types.

SingularVectors.UniversalEnvelope.AbstractElem — Type

abstract type AbstractElem{T<:Number}

Abstract type of elements in algebras.

SingularVectors.UniversalEnvelope.LieElem — Type

LieElem{T<:Number}

An element of the Lie algebra.

SingularVectors.UniversalEnvelope.SupElem — Type

SupElem{T<:Number}

LieElem type of a supalgebra.

SingularVectors.UniversalEnvelope.AbstractEnvElem — Type

abstract type AbstractEnvElem

Abstract type of elemnets in enveloping algebras.

SingularVectors.UniversalEnvelope.LieEnvElem — Type

LieEnvElem{T<:Number}

An element of the universal enveloping algebra of a Lie algebra.

SingularVectors.UniversalEnvelope.SupEnvElem — Type

SupEnvElem{T<:Number}

An element of the universal enveloping algebra of a Lie supalgebra.

Enveloping algebra of Lie algebra